Selection of coefficients based on the outcome of permutateLOO

This function assesses the original coefficients against the ones obtained from permutated samples as returned by permutateLOO() (with the parameter return_coeffs set to TRUE). It returns a vector of weights as well as a dichotomic mask reflecting which values pass a given threshold.

select_coefficients(coeffs, permutated_coeffs, threshold = 0.05)

Arguments

coeffs
Coefficients can either be passed as a list (as one created, for example, by FCnetLOO()), or as a data.frame (which must, however, have the same structure as the one returned by FCnetLOO()). This function finds consensus coefficients first trough the function specified by optionsFCnet("consensus_function"),which by default is the median.

permutated_coeffs
A list, such as an object created by permutateLOO() (with the parameter return_coeffs set to TRUE), or a data.frame (which must, however, have the same structure as the one returned by permutateLOO())

threshold
Significance threshold against which evaluate the coefficients. It is assumed to be two-tailed, and defaults to 0.05.

Value

A list with two entries: coeffs, i.e. the values of coefficients passing the statistical threshold, and mask, reporting which coefficients are retained.